backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
redis==5.0.1
websockets==12.0
python-multipart==0.0.6
textblob==0.17.1
language-tool-python==2.7.1